requirements.txt
fastapi
uvicorn
strawberry-graphql
boto3
python-dotenv
pymongo
